@@ -4,13 +4,13 @@ description: Grinder tasks for releasing Dart CLI packages.
 homepage: https://github.com/google/dart_cli_pkg
 
 environment:
-  sdk: '>=2.7.0 <3.0.0'
+  sdk: '>=2.12.0-0 <3.0.0'
 
 dependencies:
   archive: ^3.0.0-nullsafety
   async: ^2.5.0-nullsafety
   charcode: ^1.2.0-nullsafety
-  collection: ^1.15.0-nullsafety
+  collection: ^1.15.0-nullsafety.4
   crypto: ^2.2.0-nullsafety
   grinder: ^0.8.7-nullsafety
   http: ^0.13.0-nullsafety
