@@ -38,7 +38,7 @@ final githubRepo = InternalConfigVariable.fn<String>(() =>
 
 /// Returns the GitHub repo name from the Git configuration's
 /// `remote.origin.url` field.
-String _repoFromOrigin() {
+String? _repoFromOrigin() {
   try {
     var result = Process.runSync("git", ["config", "remote.origin.url"]);
     if (result.exitCode != 0) return null;
@@ -52,7 +52,7 @@ String _repoFromOrigin() {
 /// Parses a GitHub repo name from an SSH reference or a `git://` URL.
 ///
 /// Returns `null` if it couldn't be parsed.
-String _parseGit(String url) {
+String? _parseGit(String url) {
   var match = RegExp(r"^(git@github\.com:|git://github\.com/)"
           r"(?<repo>[^/]+/[^/]+?)(\.git)?$")
       .firstMatch(url);
@@ -62,7 +62,7 @@ String _parseGit(String url) {
 /// Parses a GitHub repo name from an HTTP or HTTPS URL.
 ///
 /// Returns `null` if it couldn't be parsed.
-String _parseHttp(String url) {
+String? _parseHttp(String url) {
   var match = RegExp(r"^https?://github\.com/([^/]+/[^/]+?)(\.git)?($|/)")
       .firstMatch(url);
   return match == null ? null : match[1];
@@ -107,7 +107,7 @@ String get _authorization =>
 ///
 /// If this is set to `null`, or by default if no `CHANGELOG.md` exists, no
 /// release notes will be added to the GitHub release.
-final githubReleaseNotes = InternalConfigVariable.fn<String>(() {
+final githubReleaseNotes = InternalConfigVariable.fn<String?>(() {
   if (!File("CHANGELOG.md").existsSync()) return null;
 
   return _lastChangelogSection() +
