@@ -32,7 +32,7 @@ import 'info.dart';
 /// The raw YAML of the pubspec.
 final rawPubspec =
     loadYaml(File('pubspec.yaml').readAsStringSync(), sourceUrl: Uri(path: 'pubspec.yaml'))
-        as Map<Object, Object>/*!*/;
+        as Map<Object, Object>;
 
 /// The set of entrypoint paths for executables defined by this package.
 Set<String> get entrypoints => p.PathSet.of(executables.value.values);
@@ -97,9 +97,8 @@ Future<String> get license => _licenseMemo.runOnce(() async {
       // dependencies. This also includes dev dependencies, but it's possible those
       // are compiled into the distribution anyway (especially for stuff like
       // `node_preamble`).
-      // TODO: remove as
       var packageConfigUrl = await Isolate.packageConfig;
-      var packageConfig = await loadPackageConfigUri(packageConfigUrl);
+      var packageConfig = await loadPackageConfigUri(packageConfigUrl!);
 
       // Sort the dependencies alphabetically to guarantee a consistent
       // ordering.
@@ -146,7 +145,7 @@ String _readSdkLicense() {
 
 /// Returns the contents of the `LICENSE` file in [dir], with various possible
 /// filenames and extensions, or `null`.
-String _readLicense(String dir) {
+String? _readLicense(String dir) {
   if (!Directory(dir).existsSync()) return null;
 
   var possibilities = Directory(dir)
@@ -228,7 +227,7 @@ String humanOSName(String os) {
 ///
 /// This converts each element of [iter] to a string and separates them with
 /// commas and/or [conjunction] (`"and"` by default) where appropriate.
-String toSentence(Iterable<Object> iter, {String conjunction}) {
+String toSentence(Iterable<Object> iter, {String? conjunction}) {
   if (iter.length == 1) return iter.first.toString();
   conjunction ??= 'and';
   return iter.take(iter.length - 1).join(", ") + " $conjunction ${iter.last}";
@@ -325,8 +324,7 @@ Future<String> cloneOrPull(String url) async {
 /// structure of lists and arrays of immutable scalar objects).
 Object freezeJson(Object object) {
   if (object is Map<String, Object>) return freezeJsonMap(object);
-  // TODO: remove as
-  if (object is List) return List<Object>.unmodifiable(object.map(freezeJson));
+  if (object is List<Object>) return List<Object>.unmodifiable(object.map(freezeJson));
   return object;
 }
 
