@@ -62,14 +62,14 @@ void main() {
 
     test("isn't called if the value is overridden", () {
       var variable =
-          InternalConfigVariable.fn<int>(expectAsync0(() => null, count: 0));
+          InternalConfigVariable.fn<int?>(expectAsync0(() => null, count: 0));
       variable.value = 12;
       expect(variable.value, equals(12));
     });
 
     test("isn't called if the function is overridden", () {
       var variable =
-          InternalConfigVariable.fn<int>(expectAsync0(() => null, count: 0));
+          InternalConfigVariable.fn<int?>(expectAsync0(() => null, count: 0));
       variable.fn = () => 12;
       expect(variable.value, equals(12));
     });
@@ -158,7 +158,6 @@ void main() {
     group("the freeze function", () {
       test("isn't called if ConfigVariable.freeze() isn't", () {
         var variable = InternalConfigVariable.value(1,
-            // TODO: no dynamic
             freeze: expectAsync1((n) => 0, count: 0));
         expect(variable.value, equals(1));
       });
